requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alive-progress==3.3.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
black==22.12.0
blinker==1.9.0
braceexpand==0.1.7
captcha==0.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
discord-sentry-reporting==2.0.4
discord.py==2.6.4
distro==1.9.0
docstring-parser==0.15
dpytest==0.7.0
Flask==3.1.2
frozenlist==1.8.0
graphemeu==0.7.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
import_expression==2.2.1.post1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jishaku==2.6.3
jiter==0.11.1
leetscrape==1.0.1
markdownify==0.11.6
marko==1.3.1
MarkupSafe==3.0.3
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.7.0
packaging==25.0
pandas==1.5.3
pathspec==0.12.1
peewee==3.18.2
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
pydantic==1.10.24
pygit2==1.18.2
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sentry-sdk==2.42.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
